more cache-test cleanup
